+++ src/main/java/net/MrGise/mmm/network/ClientKnowledgeData.java
@@ -1,4 +1,4 @@
-package net.MrGise.mmm.resource;
+package net.MrGise.mmm.network;
 
 public class ClientKnowledgeData {
     public static boolean ALL_KNOWING = false;