+++ src/main/java/net/MrGise/mmm/network/ModNetwork.java
@@ -1,5 +1,7 @@
-package net.MrGise.mmm.resource;
+package net.MrGise.mmm.network;
 
+import net.MrGise.mmm.network.compat.TransferBowyeryRecipePacket;
+import net.MrGise.mmm.network.compat.TransferThingamajigRecipePacket;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraftforge.network.NetworkRegistry;
 import net.minecraftforge.network.simple.SimpleChannel;
@@ -23,6 +25,20 @@ public class ModNetwork {
                 SyncAllKnowingPacket::encode,
                 SyncAllKnowingPacket::new,
                 SyncAllKnowingPacket::handle
+        );
+        CHANNEL.registerMessage(
+                packetId++,
+                TransferThingamajigRecipePacket.class,
+                TransferThingamajigRecipePacket::encode,
+                TransferThingamajigRecipePacket::new,
+                TransferThingamajigRecipePacket::handle
+        );
+        CHANNEL.registerMessage(
+                packetId++,
+                TransferBowyeryRecipePacket.class,
+                TransferBowyeryRecipePacket::encode,
+                TransferBowyeryRecipePacket::new,
+                TransferBowyeryRecipePacket::handle
         );
     }
 }